feat: Images
Taken from the .docx file as .zip. All the .emf files have been converted using
LibreOffice Draw. In LibreOffice Draw the page-size has been adjusted to the
size of the content and then exported to a .svg file for display. Intermediary
step files have been added for completeness, and because it is not yet clear how
edits will be made in the future.

Signed-off-by: Nico Rikken <nico.rikken@alliander.com>