feat: bibliography links to footnotes
There is no one-to-one match of global bibliography links as in a document.
Rather than linking to the bibliography page and readers losing their position
in the document it is implemented using footnotes. This comes with a couple of
downsides:

- duplicating the reference description inside the files
- mixing existing footnotes with bibliographyic references

Still this seems the most fitting approach for now.

The footnote id's have been prefixed with a capital letter B to prevent
conflicts while keeping the same number as currently present in the document.

It makes sense to take another look in this in the future to perhaps remove the
global bibliography and to prefer inline links.

Signed-off-by: Nico Rikken <nico.rikken@alliander.com>