chore: back to development version
Back to development version, as the github releases will be labeld the 'latest'
version already. So for example 'latest' will be an alias for v3.2.0 if v3.2.0
and v3.1.0 have been published.

To avoid confusion the override bar with the notification about not being on the
latest release is disabled for now.

This pull-request is best complemented by a commit to the `gh-pages` branch to
remove the publish under the name 'latest'.

Signed-off-by: Nico Rikken <nico.rikken@alliander.com>
Signed-off-by: Daniel Wich <info@capacious.nl>